It works!
This is a very tenuous use of the phrase "working". There  is a working
example of generating lists of partials in `queue_example.py` and `model.py`.

There are still bugs in that generating names doesn't work perfectly
(redundancy in naming) and the actual testing of properties applied to
models does not happen yet. Finally, there are plenty of debug
statements and useless code to be removed/refactored.